move device to its own file. make important arguments @required, update docs
