+++ packages/golden_toolkit/lib/golden_toolkit.dart
@@ -7,6 +7,7 @@
 /// ***************************************************
 library golden_toolkit;
 
+export 'src/device.dart';
 export 'src/font_loader.dart';
 export 'src/golden_builder.dart';
 export 'src/multi_screen_golden.dart';
