+++ packages/golden_toolkit/lib/src/multi_screen_golden.dart
@@ -7,6 +7,7 @@
 /// ***************************************************
 import 'dart:ui';
 import 'package:flutter_test/flutter_test.dart';
+import 'device.dart';
 import 'testing_tools.dart';
 
 Future<void> _onlyPumpAndSettle(WidgetTester tester) async =>
@@ -55,57 +56,6 @@ Future<void> multiScreenGolden(
       '$goldenFileName.${device.name}',
       customPump: customPump,
       skip: skip,
-    );
-  }
-}
-
-/// This [Device] is a configuration for golden test. Can be provided for [multiScreenGolden]
-
-class Device {
-  /// This [Device] is a configuration for golden test. Can be provided for [multiScreenGolden]
-  const Device({
-    this.size,
-    this.devicePixelRatio = 1.0,
-    this.name,
-    this.textScale = 1.0,
-  });
-
-  /// [phone] one of the smallest phone screens
-  static const Device phone = Device(name: 'phone', size: Size(375, 667));
-
-  /// [tabletLandscape] example of tablet that in landscape mode
-  static const Device tabletLandscape =
-      Device(name: 'tablet_landscape', size: Size(1366, 1024));
-
-  /// [tabletPortrait] example of tablet that in portrait mode
-  static const Device tabletPortrait =
-      Device(name: 'tablet_portrait', size: Size(1024, 1366));
-
-  /// [name] specify device name. Ex: Phone, Tablet, Watch
-
-  final String name;
-
-  /// [size] specify device screen size. Ex: Size(1366, 1024))
-  final Size size;
-
-  /// [devicePixelRatio] specify device Pixel Ratio
-  final double devicePixelRatio;
-
-  /// [textScale] specify custom text scale
-  final double textScale;
-
-  /// [copyWith] convenience function for [Device] modification
-  Device copyWith({
-    Size size,
-    double devicePixelRatio,
-    String name,
-    double textScale,
-  }) {
-    return Device(
-      size: size ?? this.size,
-      devicePixelRatio: devicePixelRatio ?? this.devicePixelRatio,
-      name: name ?? this.name,
-      textScale: textScale ?? this.textScale,
     );
   }
 }
